Maintain map list instead of numeric map identification, add map extension, update player/enemy_killed events, update enemy death handling

+++ battlecity/controller.cpp
@@ -4,9 +4,13 @@
 
 #include "ecs/systems.h"
 
+#include <QDir>
 #include <QFileInfo>
 
-static constexpr auto map_name_pattern = ":/maps/map_%1";
+static constexpr auto map_extension = "bsmap";
+static constexpr auto map_dir_path = ":/maps/";
+
+static const uint32_t map_switch_pause_duration{ 3000 };
 
 namespace game
 {
@@ -21,6 +25,15 @@ controller::controller( const game_settings& settings, ecs::world& world ) :
 
 void controller::init()
 {
+    QDir map_dir{ map_dir_path };
+    map_dir.setNameFilters( QStringList{} <<  QString{ "*.%1" }.arg( map_extension ) );
+    m_levels = map_dir.entryList( QDir::NoDotAndDotDot | QDir::AllEntries );
+
+    if( m_levels.empty() )
+    {
+        throw std::logic_error{ "No maps found" };
+    }
+
     // create systems
     std::unique_ptr< ecs::system > move_system{ new system::movement_system{ m_world } };
 
@@ -29,9 +42,9 @@ void controller::init()
 
     std::unique_ptr< ecs::system > proj_system{
         new system::projectile_system{ m_settings.get_projectile_size(),
-                                       m_settings.get_projectile_damage(),
-                                       m_settings.get_projectile_speed(),
-                                       m_world } };
+                    m_settings.get_projectile_damage(),
+                    m_settings.get_projectile_speed(),
+                    m_world } };
 
     std::unique_ptr< ecs::system > respawn_system{
         new system::respawn_system{ std::chrono::milliseconds{ m_settings.get_respawn_delay_ms() },
@@ -40,7 +53,7 @@ void controller::init()
     std::unique_ptr< ecs::system > tank_ai_system{
         new system::tank_ai_system{ m_settings.get_ai_chance_to_fire(), m_world } };
 
-    load_level( m_level );
+    load_level();
     m_need_to_load_level = false;
 
     m_world.add_system( *move_system );
@@ -58,27 +71,23 @@ void controller::init()
     m_world.subscribe< event::level_completed >( *this );
 }
 
-bool controller::load_level( uint32_t level )
+QString get_map_path( const QString& map_name )
 {
-    QString map_path{ QString{ map_name_pattern }.arg( level ) };
-    QFileInfo map_file{ map_path };
-    bool map_valid{ map_file.exists() && map_file.isFile() };
+    return QString{ map_dir_path } + map_name;
+}
 
-    if( map_valid )
+void controller::load_level()
+{
+    read_map_file( m_map_data,
+                   get_map_path( m_levels.front() ),
+                   m_settings,
+                   m_world,
+                   m_mediator );
+
+    for( auto& system : m_systems )
     {
-        read_map_file( m_map_data,
-                       QString{ map_name_pattern }.arg( level ),
-                       m_settings,
-                       m_world,
-                       m_mediator );
-
-        for( auto& system : m_systems )
-        {
-            system->init();
-        }
+        system->init();
     }
-
-    return map_valid;
 }
 
 void controller::start()
@@ -86,7 +95,7 @@ void controller::start()
     m_tick_timer->start( 1000 / m_settings.get_fps() );
     if( m_mediator )
     {
-        m_mediator->level_started( m_level );
+        m_mediator->level_started( m_map_data.get_map_name() );
     }
 }
 
@@ -120,29 +129,42 @@ int controller::get_tile_height() const noexcept
     return m_settings.get_tile_size().height();
 }
 
-uint32_t controller::get_level() const noexcept
+const QString& controller::get_level() const noexcept
 {
-    return m_level;
+    return m_map_data.get_map_name();
 }
 
 uint32_t controller::get_remaining_frag_count()
 {
-     component::level_info* l = m_world.get_components< component::level_info >().front();
-     return l->get_kills_to_win() - l->get_player_kills();
+    component::level_info* l = m_world.get_components< component::level_info >().front();
+    return l->get_kills_to_win() - l->get_player_kills();
+}
+
+uint32_t controller::get_map_switch_pause_duration() const noexcept
+{
+    return map_switch_pause_duration;
 }
 
 void controller::on_event( const event::level_completed& event )
 {
     // TODO: check if level is present
-     m_need_to_load_level = true;
+    m_need_to_load_level = true;
     if( event.get_result() == level_game_result::victory )
     {
-        ++m_level;
+        m_levels.pop_front();
     }
 
     if( m_mediator )
     {
-        m_mediator->level_ended( event.get_result() );
+        if( !m_levels.empty() ) // all maps have been beaten
+        {
+            m_mediator->level_ended( event.get_result() );
+        }
+        else
+        {
+            stop();
+            m_mediator->game_ended( level_game_result::victory );
+        }
     }
 }
 
@@ -150,28 +172,22 @@ void controller::tick()
 {
     if( m_need_to_load_level )
     {
-        if( m_mediator )
-        {
-            m_mediator->remove_all();
-        }
+        m_need_to_load_level = false;
 
-        m_world.reset();
-        if( load_level( m_level ) )
+        if( !m_levels.empty() )
         {
+            std::this_thread::sleep_for( std::chrono::milliseconds{ map_switch_pause_duration } );
             if( m_mediator )
             {
-                m_mediator->level_started( m_level );
+                m_mediator->remove_all();
             }
 
-            m_need_to_load_level = false;
-        }
-        else // all maps have been beaten
-        {
+            m_world.reset();
+            load_level();
+
             if( m_mediator )
             {
-                m_mediator->game_ended( level_game_result::victory );
-                stop();
-                return;
+                m_mediator->level_started( m_map_data.get_map_name() );
             }
         }
     }
